requirements.txt
numpy>=1.24
playwright>=1.40
Pillow>=10.0

# RL Benchmark dependencies
gymnasium>=0.29
torch>=2.0
matplotlib>=3.7
pandas>=2.0
